requirements.txt
fastapi
uvicorn
pandas
python-multipart
openai
python-pptx